server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gtts
moviepy==1.0.3
google-generativeai
elevenlabs